move files to respective directories
