+++ src/App.js
@@ -1,10 +1,10 @@
-import React from 'react';
-import './App.css';
+import React from "react";
+import "./css/App.css";
 
 function App() {
   return (
     <div className="App">
-      <h1>Fast Fingers </h1>
+      <h1>Fast Fingers</h1>
     </div>
   );
 }
