+++ src/css/App.css
@@ -1,0 +1,1 @@
+/* No CSS */
